Begin updates to v20 of Essentials
As it turns out, v20 came out a whole 2 months ago. Whoops. Well this is the first commit of *many* to the road to v20. to start here's a few small changes that I've made so far.

The way text is drawn has changed slightly, seemingly the inherent padding was done away with, so no need to fiddle with that anymore.

that being said, I still had some issues with displaying some of the list text so I do have a small adjustment, will have to see if that can be done away with as well.

I noticed while I was working on that that I had a file dependency that I really shouldn't have... so I removed it! now you can freely remove OAK.png from the files if you aren't using it, without breaking everything else.

There are still some problems to be fixed, such as scrolling the name list, and I don't even want to know if my Infobox stuff broke, but I'll get to it soon enough.

Internal logic for the name list seems to still work, it's just the repositioning of the list that is broken, so shouldn't be that hard to fix.

While not noted directly, the PBS changes didn't impact the program in any way directly, but did give me an idea.