Finally finished creating most of what is needed for compiling data
Compiling data from a PBS file is DONE. That being said, there is still some other stuff that needs to be handled which I'm not currently doing. But that's for another update. Compiling data is complete, and working, and I also figured out conversion from old save to new. Still have a small bit left to do though