Added the ability to change which folder the PBS file is in
This small update adds in another option within the userFunctions.rb file to allow the user to change the folder location of their PBS file. In addition to that, I modified the code within ensure CompiledData to always load @@compiledData and removed the slightly annoying error message that would appear when loading up the plugin for the first time.
The changes to CompileData.rb corresponds to allowing the user to change the PBS location, changing every usage of "PBS/FameChecker" to COMPILE_FOLDER.