Fix save parity again...
Fixed an error that would cause a desync across save files if fame info elements were removed or added anywhere except the end, if elements were removed and added at the same time, or if elements were swapped. unfortunately, if a save file was already broken, this update doesn't save it.
Additionally, I added the ability to change the display name of famous people, fixed a small bug where the first info element would still animate when still at the famous person select menu, added the ability to add the item to the favorites menu, and allowed the usage of Strings within the  user functions instead of Symbols.