Compiler and Save Parity Fixes
Fixed an issue with the save parity system in which the compiled data's Complete values would also get modified, This bug wouldn't have ever affected anybody.
Updated the compiler code a little bit, separated the compilation function into two functions, one for the famous people, and another for the fame info.
Added a new system that can generate a fame_targets.txt file from either the compiled data or default data, and hooked that into the compile code so that it automatically generates a new fame_targets.txt if one doesn't exist.