Added the final missing element from the actual games
I missed a small but significant detail from the fame checker within Fire Red and Leaf Green, When the player has found all pieces of info for a famous person, there is supposed to be a letter from that famous person addressed to them, I didn't have that because I was testing on a non-completed copy of Fire Red. My bad. This is now implemented, both in game and in the debug menu. It uses the element :PlayerMessage.